Moved the "JwtToken" and the "RegistryClient" classes to the lib directory
They were previously inside the "app/models" directory, which does not make
much sense since they're not models at all. Instead of this, it's better to
have them inside of the "lib" directory.

Moreover, since the JwtToken serializes into the `encoded_token` method, and
it's only used once, I prefer to just render the appropiate hash for it. This
class will undertake some refactoring later on.
