--- app/serializers/jwt_token_serializer.rb
@@ -1,7 +1,0 @@
-class JwtTokenSerializer < ActiveModel::Serializer
-  def token
-    object.encoded_token
-  end
-
-  attributes :token
-end
